Update initial app name across the repo
Why these changes are being introduced:
"app" was too generic a term to use as the initial app name. It meant a
find-and-replace across the repo would change many more things than the
actual app name (for example, documentation including the words "app" or
"application"). A better initial name is one that is not part of other
words and would *only* be used in places where the app name matters in
the code, making it easy to find and replace.

How this addresses that need:
* Replaces the app name "app" with "my_app" across the repo to ensure
  future users of this template can easily update the app name when
  first creating it.