Initial scaffolding of CLI app
Why these changes are being introduced:

Scaffolds the browsertrix-harvester app.  This harvester requires a Docker container for web crawls,
and this commit sets up the python click CLI interface, plus harvester, as part of a single
Docker container.

Background:

This is the first PR in a series of PRs to build iteratively build the app and
provide code review.

Relevant ticket(s):

https://mitlibraries.atlassian.net/browse/TIMX-247