Add web crawling functionality
Why these changes are being introduced:

The first step in a harvest for this application is performing a web crawl.  This commit adds a Crawler class and a CLI command
for configuring and running browsertrix web crawls inside a Docker container.

At this point, no metadata is parsed, but a WACZ file is produced that will eventually support parsing of metadata records for
websites crawled.

How this addresses that need:

  * adds Crawler class that wraps the functionality of browsertrix-crawler
  * adds CLI commands and configurations for running crawl
  * wraps a web crawl under a larger "harvest" umbrella for eventual metadata parsing as well

Side effects of this change:

None

Relevant ticket(s):
https://mitlibraries.atlassian.net/browse/TIMX-247