Refactor metadata record parsing from WACZ helpers
Why these changes are being introduced:

Previously, much of this functionality was all grouped under a single CrawlParser class.  This was
mixing the task of generating metadata records from a completed crawl, and the somewhat complex
helpers and utilities for interacting with a WACZ file.

How this addresses that need:

This refactor breaks the functionality into two classes: CrawlMetadataParser and WACZClient.  This
helps isolate the functionality, allowing for easier code readability and testing.

Additionally, another pass at variable renaming and docstring improvement.