Support JSONLines as output format
Why these changes are being introduced:

When this harvester was first created, Transmogrifier would only accept XML as input.
With the creation of the GeoHarvester, Transmog now supports JSONLines as an input
format.  Writing JSONLines is a handy option for this harvester as well.

How this addresses that need:
* Write JSONLines if a `.jsonl` output file is passed.  Even simpler than the XML
output, we already have the data in a dataframe that converts nicely to
JSONLines.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-542