Updates For New Shared Workflows
Why these changes are being introduced:
This app is undergoing regular maintenance, so we are adding the change
to the new Makefile commands and the new workflows as part of this
maintenance.

How this addresses that need:
* Update Makefile with outputs from changes to mitlib-tf-workloads-ecr
* Update the dev-build workflow with the output from changes to
mitlib-tf-workloads-ecr
* Update the stage-build workflow with the output from changes to
mitlib-tf-workloads-ecr
* Update the prod-promote workflow with the output from changes to
mitlib-tf-workloads-ecr
* Update the trigger for the CI workflow to only run for PRs
* Update .gitignore to skip the temporary .arch_tag file that is
created (and sometimes not deleted) by the make commands that build
the container

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IN-1500