Surface crawl stats for non-verbose logging
Why these changes are being introduced:

With verbose logging, we see most all output from the
harvester via stdout and stderr from the subprocess process.
This is overkill when we might only want to know occassionally
how far along the crawl is.

How this addresses that need:
We look for a harvester log line with "crawlStatus" that,
when parsed, contains information about how far along the
crawl is.  Instead of logging this each time we encounter it,
we buffer it and default to only showing every 100th instance we
see.  It's not super precise, but roughly logs every 100 websites
crawled.

Side effects of this change:
* Crawl status every 100th line for non-verbose logging.

Relevant ticket(s):
* None