Handle crawls with different pages and CDX data
Why these changes are being introduced:

After performing some crawls with an explicit list of seed URLs,
it was noted that the /pages and /indexes data was different in
the final output of the crawl. This broke some metadata parsing
assumptions we were making about files.

How this addresses that need:

These updates allow for different files, with different contents,
given the crawl type.  In short, we now can handle a non-compresseed
CDX file, and better handle if pages.jsonl or extraPages.jsonl is
empty.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/TIMX-562