Support pre-crawl, manual sitemap parsing
Why these changes are being introduced:

The browsertrix harvester has sitemap support, but it's tightly
coupled with the crawl itself.  For example, sitemaps are parsed and URLs extracted
as they are discovered  through crawling.  While they can be defined in advance, even
this has some non-deterministic behavior because the crawl begins before
they are all parsed.

Additionally, even when sitemaps are used to seed the crawl, a concrete, unchanging
list of URLs discovered from the sitemaps is not available after the crawl.

How this addresses that need:

An optional, pre-crawl step is added where we manually parse sitemaps
passed as CLI arguments using a standalone python library.  With the
sitemaps parsed, and URLs discovered, we optionally also write the list
of URLs to a file independant of the crawl itself.

For the crawl, this same list of URLs is used and passed to the crawler
via --urlFile option, ensuring that explicit list of URLs and *only* those URLs
are crawled (if no other seeds are passed).

This sets us up for future work of our harvester analyzing previous
URL text files to look for URL diffs, indicating new or deleted URLs.

Side effects of this change:
* Additional CLI arguments, likely driving how this harvester will
be used in production.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-93