Rework from / to dates in sitemap parsing
Why these changes are being introduced:

From and To dates are optional when pre-crawl parsing sitemaps.
When passed, they should affect the URLs used for the crawl itself,
but should NOT be used for the file written that is intended
to serve as a snapshot of all URLs discovered.

How this addresses that need:

Override sitemap_from_date and sitemap_to_date during the .write()
method during external file write.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-93