Generate delete metadata records
Why these changes are being introduced:

Ideally, when running in a daily / incremental capacity, TIMDEX harvester
can return metadata records about records that have been deleted in the
source system.  For this harvester, the source system are webesites.

We needed a way for this harvester to say that given a harvest today,
pages X,Y,Z have been deleted since the last harvest.

How this addresses that need:

A new CLI argument is added, `--previous-sitemap-urls-file`, which
indicates to the harvester to try and download that list of URLs from
a previous manual, pre-crawl sitemap parsing and compare against the
URLs from the current crawl.  Any URLs that were present before, but
are not now, are considered deleted.

A new field is also added to the output metadata records,
status=active|deleted.

Once Transmogrifier is updated to key on this status field, this
harvester will be capable of communicating deletions in TIMDEX.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-97