Streamline sitemap CLI arg
Why these changes are being introduced:

The first pass at manually passing sitemaps to the CLI included
a sitemap "root" and sitemap "paths", which closely mirrored the
python library that we use for pre-crawl sitemap parsing. This had
two major drawbacks:

1. You could really only provide sitemaps for a single domain.
2. It was exposing too much of the underyling logic to the
CLI interface.

At the CLI level, you should just be able to repeat a `--sitemap`
argument as many times as needed.

How this addresses that need:

* remove the `--parse-sitemaps-pre-crawl` CLI arg, and default to this
when providing `--sitemap`
* reduce down to just repeatable `--sitemap`
* during parsing, a bit more logic to detect different domains
and group them for passing to python library for parsing

Side effects of this change:
* Able to pass sitemaps from multiple domains
* CLI args are simpler

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-93