Remove crawler workers defaults
Why these changes are being introduced:

With the CLI args and Crawler class providing defaults for the number
of workers a crawl will use, this prevented setting this in a configuration
YAML file which was clobbered by those settings.

How this addresses that need:

Removing the CLI arg and Crawler class defaults allows a configuration YAML file
to provide a 'workers' property that will get used.

Side effects of this change:

When the pipeline lambda builds a CLI command for this harvester, it need not
set '--num-workers', instead relying on the configuration YAML to contain
this value.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-86