Rework harvester to return "records" vs "metadata records"
Why these changes are being introduced:

A decision was made to pivot the output of this harvester from "metadata records" to
just "records".  Due to the highly varying structure of websites, any attempt to parse
metadata about the website -- e.g. using OpenGraph or Dublin Core tags in the head
section -- was inherently opinionated towards the particular sites getting crawled.  This
was leaning into an unsustainable pattern that when we encountered new websites, with
metadata in different places, we'd have to hardcode that metadata extraction into the
harvester such that it would be available downstream.

Instead of the harvester returning metadatda records about the websites captured in the
crawl, the decision has the harvester returning the full HTML for each website as part
of the "record" that is written as output.  This HTML can then be used in downstream
applications like Transmogrifier to extract metadata, which are designed to be
opinionated about a particular source.

The implicit proposal here is that parsing metadata from messy HTML is different than
parsing metadata from structured data like an EAD or METS file, but it's still the
work of extracting *TIMDEX* metadata from some kind of source data.

How this addresses that need:

There are three major changes here:

1. Change "metadata" language to "records" through the codebase.  This is needed to
convey that we aren't really creating "metadata" records per se, but instead returning
structured "records" with the HTML content of each website.  The harvester is still
doing important work on top of the raw web crawl, just not parsing metadata.

2. Include the full HTML of the website in the output record.  Because HTML can have
all kinds of problematic characters, we base64 ASCII encode it, and store in a field
called `base64_html`.

3. We remove any CLI, class, or method arguments around including fulltext or
extracting keywords.  Given that we are now including the full HTML in the output,
which contains the fulltext of the website in a form we have more control over parsing,
we don't need the harvester to extract fulltext as part of the metadatda record.  This
also fully completes the removal of keyword extraction which were somewhat experimental
and will likely be improved upon by embeddings.

Side effects of this change:
* The browsertrix-harvester now returns "records" vs "metadata records" as its primary
output for each run.  The records contain some light metadatda about the website itself,
e.g. URL, etc., but an equally important column is the full, rendered HTML of the page.
It is expected that downstream systems like Transmogrifier will be responsible for
extracting metadata in a more focused and opinionated fashion.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-258