Add response headers to output records
Why these changes are being introduced:

Now that we include the full, rendered HTML in output records for downstream systems
to parse and use, it would be helpful to also have the associated HTTP response headers
which provide additional information.  Sometimes these can help indicate when the website
was updated, or alternate links, etc.  It may very well not get used, but it would be
beneficial to err on the side of including.

How this addresses that need:
* During record generation, include the response headers as new `response_headers`
column.

Side effects of this change:
* None

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/USE-272